+++ packages/react-querybuilder/src/utils/formatQuery/formatQuery.ts
@@ -15,6 +15,7 @@ import type {
   RuleProcessor,
   RuleType,
   RuleValidator,
+  SQLPreset,
   ValidationMap,
   ValidationResult,
 } from '../../types/index.noReact';
@@ -39,6 +40,22 @@ import {
   numerifyValues,
   quoteFieldNamesWithArray,
 } from './utils';
+
+const sqlDialectPresets = {
+  ansi: {},
+  sqlite: { paramsKeepPrefix: true },
+  oracle: {},
+  mssql: {
+    quoteFieldNamesWith: ['[', ']'],
+    concatOperator: '+',
+  },
+  mysql: { concatOperator: 'CONCAT' },
+  postgresql: {
+    quoteFieldNamesWith: '"',
+    numberedParams: true,
+    paramPrefix: '$',
+  },
+} satisfies Record<SQLPreset, FormatQueryOptions>;
 
 /**
  * Generates a formatted (indented two spaces) JSON string from a query object.
@@ -155,8 +172,12 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
       ruleProcessorInternal = defaultRuleProcessorJSONata;
     }
   } else {
-    format = (options.format ?? 'json').toLowerCase() as ExportFormat;
-    const { valueProcessor = null, ruleProcessor = null } = options;
+    const optionsWithPresets = {
+      ...(sqlDialectPresets[options.preset ?? 'ansi'] ?? {}),
+      ...options,
+    };
+    format = (optionsWithPresets.format ?? 'json').toLowerCase() as ExportFormat;
+    const { valueProcessor = null, ruleProcessor = null } = optionsWithPresets;
     if (typeof ruleProcessor === 'function') {
       ruleProcessorInternal = ruleProcessor;
     }
@@ -179,18 +200,19 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
                   : format === 'jsonata'
                     ? (ruleProcessorInternal ?? defaultRuleProcessorJSONata)
                     : defaultValueProcessorByRule;
-    quoteFieldNamesWith = quoteFieldNamesWithArray(options.quoteFieldNamesWith);
-    validator = options.validator ?? (() => true);
-    fields = toFullOptionList(options.fields ?? []);
-    fallbackExpression = options.fallbackExpression ?? '';
-    paramPrefix = options.paramPrefix ?? ':';
-    paramsKeepPrefix = !!options.paramsKeepPrefix;
-    numberedParams = !!options.numberedParams;
-    parseNumbers = !!options.parseNumbers;
-    placeholderFieldName = options.placeholderFieldName ?? defaultPlaceholderFieldName;
-    placeholderOperatorName = options.placeholderOperatorName ?? defaultPlaceholderOperatorName;
-    quoteValuesWith = options.quoteValuesWith ?? "'";
-    concatOperator = options.concatOperator ?? '||';
+    quoteFieldNamesWith = quoteFieldNamesWithArray(optionsWithPresets.quoteFieldNamesWith);
+    validator = optionsWithPresets.validator ?? (() => true);
+    fields = toFullOptionList(optionsWithPresets.fields ?? []);
+    fallbackExpression = optionsWithPresets.fallbackExpression ?? '';
+    paramPrefix = optionsWithPresets.paramPrefix ?? ':';
+    paramsKeepPrefix = !!optionsWithPresets.paramsKeepPrefix;
+    numberedParams = !!optionsWithPresets.numberedParams;
+    parseNumbers = !!optionsWithPresets.parseNumbers;
+    placeholderFieldName = optionsWithPresets.placeholderFieldName ?? defaultPlaceholderFieldName;
+    placeholderOperatorName =
+      optionsWithPresets.placeholderOperatorName ?? defaultPlaceholderOperatorName;
+    quoteValuesWith = optionsWithPresets.quoteValuesWith ?? "'";
+    concatOperator = optionsWithPresets.concatOperator ?? '||';
   }
   if (!fallbackExpression) {
     fallbackExpression =